requirements.txt
Flask
flask-cors
python-dotenv
openai
langchain
langchain-openai
requests
pymongo
dnspython